+++ core/src/main/java/com/graphhopper/routing/ev/MtbRating.java
@@ -4,6 +4,6 @@ public class MtbRating {
     public static final String KEY = "mtb_rating";
 
     public static IntEncodedValue create() {
-        return new UnsignedIntEncodedValue(KEY, 3, false);
+        return new IntEncodedValueImpl(KEY, 3, false);
     }
 }