+++ core/src/test/java/com/graphhopper/routing/ev/DecimalEncodedValueTest.java
@@ -13,7 +13,7 @@ public class DecimalEncodedValueTest {
 
     @Test
     public void testInit() {
-        DecimalEncodedValue prop = new UnsignedDecimalEncodedValue("test", 10, 2, false);
+        DecimalEncodedValue prop = new DecimalEncodedValueImpl("test", 10, 2, false);
         prop.init(new EncodedValue.InitializerConfig());
         IntsRef ref = new IntsRef(1);
         prop.setDecimal(false, ref, 10d);
@@ -32,7 +32,7 @@ public class DecimalEncodedValueTest {
         IntsRef flags = carEncoder.handleWayTags(em.createEdgeFlags(), way);
         assertEquals(101.5, carAverageSpeedEnc.getDecimal(true, flags), 1e-1);
 
-        DecimalEncodedValue instance1 = new UnsignedDecimalEncodedValue("test1", 8, 0.5, false);
+        DecimalEncodedValue instance1 = new DecimalEncodedValueImpl("test1", 8, 0.5, false);
         instance1.init(new EncodedValue.InitializerConfig());
         flags = em.createEdgeFlags();
         instance1.setDecimal(false, flags, 100d);
@@ -41,7 +41,7 @@ public class DecimalEncodedValueTest {
 
     @Test
     public void testNegativeBounds() {
-        DecimalEncodedValue prop = new UnsignedDecimalEncodedValue("test", 10, 5, false);
+        DecimalEncodedValue prop = new DecimalEncodedValueImpl("test", 10, 5, false);
         prop.init(new EncodedValue.InitializerConfig());
         try {
             prop.setDecimal(false, new IntsRef(1), -1);