+++ core/src/test/java/com/graphhopper/routing/ev/DefaultEncodedValueFactoryTest.java
@@ -18,7 +18,7 @@ public class DefaultEncodedValueFactoryTest {
     @Test
     public void loadCarMaxSpeed() {
         EncodedValue enc = MaxSpeed.create();
-        UnsignedDecimalEncodedValue loadedEnc = (UnsignedDecimalEncodedValue) factory.create(enc.toString());
+        DecimalEncodedValueImpl loadedEnc = (DecimalEncodedValueImpl) factory.create(enc.toString());
         assertEquals(loadedEnc, enc);
     }